add component utils
